Moved examples to subfolders
