+++ examples/general/SMAC4BO/SMAC4BO_rosenbrock_example.py
