+++ examples/general/SMAC4HPO/SMAC4HPO_acq_rosenbrock_example.py
