+++ examples/general/SMAC4HPO/SMAC4HPO_rosenbrock_example.py
