+++ examples/general/function_minimization/fmin_rosenbrock_example.py
