+++ examples/general/function_minimization/rosenbrock_2d_delayed_func.py
